Fix style and components details
